feat: refactor user module structure; move user module and service, update imports, and implement user schema with DTOs
