+++ src/common/decorators/roles.decorator.ts
@@ -1,5 +1,5 @@
 import { SetMetadata } from '@nestjs/common';
-import { Roles } from 'src/modules/Admin/user/Schema/user.schema';
+import { Roles } from 'src/modules/user/Schema/user.schema';
 
 export const ROLES_KEY = 'roles';
 export const Role = (...roles: Roles[]) => SetMetadata(ROLES_KEY, roles);
