+++ src/modules/user/user.controller.ts
@@ -12,8 +12,7 @@ import {
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-import { AuthGuard } from '../../../common/guards/Auth.guard';
-import type { Request } from 'express';
+import { AuthGuard } from '../../common/guards/Auth.guard';
 import { Role } from 'src/common/decorators/roles.decorator';
 import { Roles } from './Schema/user.schema';
 
@@ -32,27 +31,50 @@ export class UserController {
    */
   @Post()
   create(@Body() createUserDto: CreateUserDto) {
-    console.log('Reached');
     return this.userService.create(createUserDto);
   }
 
+  /**
+   * @desc Admin get all users
+   * @access Private [Admin]
+   * @method Get
+   * @route /api/v1/user
+   */
   @Get()
   findAll() {
     return this.userService.findAll();
   }
 
+  /**
+   * @desc Admin get one user
+   * @access Private [Admin]
+   * @method Get
+   * @route /api/v1/user/:id
+   */
   @Get(':id')
   findOne(@Param('id') id: string) {
-    return this.userService.findOne(+id);
+    return this.userService.findOne(id);
   }
 
+  /**
+   * @desc Admin updates user
+   * @access Private [Admin]
+   * @method Patch
+   * @route /api/v1/user/:id
+   */
   @Patch(':id')
   update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
-    return this.userService.update(+id, updateUserDto);
+    return this.userService.update(id, updateUserDto);
   }
 
+  /**
+   * @desc Admin deletes user
+   * @access Private [Admin]
+   * @method Delete
+   * @route /api/v1/user/:id
+   */
   @Delete(':id')
   remove(@Param('id') id: string) {
-    return this.userService.remove(+id);
+    return this.userService.remove(id);
   }
 }
