+++ src/modules/user/user.module.ts
@@ -3,9 +3,8 @@ import { UserService } from './user.service';
 import { UserController } from './user.controller';
 import { MongooseModule } from '@nestjs/mongoose';
 import { User, UserSchema } from './Schema/user.schema';
-import { JWTService } from '../../jwt/jwt.service';
-import { JWTModule } from '../../jwt/jwt.module';
-import { ConfigModule, ConfigService } from '@nestjs/config';
+import { JWTService } from '../jwt/jwt.service';
+import { ConfigModule } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
 
 @Module({
